+++ tests/test_backends.py
@@ -4,9 +4,9 @@ import os
 import pytest
 
 from arca import Arca, VenvBackend, DockerBackend, Task, CurrentEnvironmentBackend
-from common import BASE_DIR, RETURN_COLORAMA_VERSION_FUNCTION, SECOND_RETURN_STR_FUNCTION, \
-    TEST_UNICODE, ARG_STR_FUNCTION, KWARG_STR_FUNCTION, WAITING_FUNCTION
 from arca.exceptions import BuildTimeoutError
+from common import BASE_DIR, RETURN_COLORAMA_VERSION_FUNCTION, SECOND_RETURN_STR_FUNCTION, \
+    TEST_UNICODE, ARG_STR_FUNCTION, KWARG_STR_FUNCTION, WAITING_FUNCTION, RETURN_STR_FUNCTION
 
 
 @pytest.mark.parametrize(
@@ -17,6 +17,9 @@ from arca.exceptions import BuildTimeoutError
     ))
 )
 def test_backends(temp_repo_func, backend, requirements_location, file_location):
+    """ Tests the basic stuff around backends, if it can install requirements from more locations,
+        launch stuff with correct cwd, works well with multiple branches, etc
+    """
     if os.environ.get("TRAVIS", False) and backend == VenvBackend:
         pytest.skip("Venv Backend doesn't work on Travis")
 
@@ -86,6 +89,41 @@ def test_backends(temp_repo_func, backend, requirements_location, file_location)
 
     assert arca.run(temp_repo_func.url, temp_repo_func.branch, task).output == "0.3.8"
 
+    # cleanup
+
+    if isinstance(backend, CurrentEnvironmentBackend):
+        backend._uninstall("colorama")
+
+    with pytest.raises(ModuleNotFoundError):
+        import colorama  # noqa
+
+
+@pytest.mark.parametrize(
+    "backend",
+    [CurrentEnvironmentBackend, VenvBackend, DockerBackend]
+)
+def test_advanced_backends(temp_repo_func, backend):
+    """ Tests the more time-intensive stuff, like timeouts or arguments,
+        things multiple for runs with different arguments are not neccessary
+    """
+    if os.environ.get("TRAVIS", False) and backend == VenvBackend:
+        pytest.skip("Venv Backend doesn't work on Travis")
+
+    kwargs = {}
+
+    if backend == DockerBackend:
+        kwargs["disable_pull"] = True
+    if backend == CurrentEnvironmentBackend:
+        kwargs["current_environment_requirements"] = None
+        kwargs["requirements_strategy"] = "install_extra"
+
+    backend = backend(verbosity=2, **kwargs)
+
+    arca = Arca(backend=backend, base_dir=BASE_DIR)
+
+    filepath = temp_repo_func.file_path
+    requirements_path = temp_repo_func.repo_path / backend.requirements_location
+
     filepath.write_text(ARG_STR_FUNCTION)
     temp_repo_func.repo.index.add([str(filepath)])
     temp_repo_func.repo.index.commit("Argument function")
@@ -104,6 +142,7 @@ def test_backends(temp_repo_func, backend, requirements_location, file_location)
         kwargs={"kwarg": TEST_UNICODE}
     )).output == TEST_UNICODE[::-1]
 
+    # test task timeout
     filepath.write_text(WAITING_FUNCTION)
     temp_repo_func.repo.index.add([str(filepath)])
     temp_repo_func.repo.index.commit("Waiting function")
@@ -112,12 +151,19 @@ def test_backends(temp_repo_func, backend, requirements_location, file_location)
     task_3_seconds = Task("test_file:return_str_function", timeout=3)
 
     with pytest.raises(BuildTimeoutError):
-        assert arca.run(temp_repo_func.url, temp_repo_func.branch, task_1_second).output == "Some string"
+        arca.run(temp_repo_func.url, temp_repo_func.branch, task_1_second)
 
     assert arca.run(temp_repo_func.url, temp_repo_func.branch, task_3_seconds).output == "Some string"
 
-    if isinstance(backend, CurrentEnvironmentBackend):
-        backend._uninstall("colorama")
+    # test requirements timeout
+    requirements_path.write_text("scipy")
 
-    with pytest.raises(ModuleNotFoundError):
-        import colorama  # noqa
+    filepath.write_text(RETURN_STR_FUNCTION)
+
+    temp_repo_func.repo.index.add([str(filepath), str(requirements_path)])
+    temp_repo_func.repo.index.commit("Updated requirements to something that takes > 1 second to install")
+
+    arca.backend.requirements_timeout = 1
+
+    with pytest.raises(BuildTimeoutError):
+        arca.run(temp_repo_func.url, temp_repo_func.branch, Task("test_file:return_str_function"))